# architecture/OpenApi/api.yaml
openapi: 3.0.0
info:
  description: >-
    This is the API for a course project at ITMO University. The theme of the
    course is "The goblin slayer contract system". Server on github.
  version: "1.1.0"
  title: Goblin Slayer Server API
servers:
  - description: Server
    url: 'http://paulrozhkin.ru:3464/'
...
paths:
...
  /api/v1/admin/users/{id}:
    # Обновление модели пользователя
    put:
      tags:
        - admin
      operationId: updateUserById
      summary: Update the user's model
      security:
        - bearerAuth: []
      parameters:
        - name: id
          description: 'User ID.'
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userUpdatableModel'
        required: true
      responses:
        '200':
          description: The user's model was updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '400':
          description: Bad Reaquest.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
  /api/v1/users/:
    #Создание новой модели пользователя
    post:
      tags:
        - users
      operationId: createUser
      summary: Create the user's model
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userCreatableModel'
        required: true
      responses:
        '200':
          description: The user's model was created successfully.
          content:
            application-json:
              schema:
                $ref: '#/components/schemas/user'
        '400':
          description: Bad Reaquest.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
  /api/v1/users/{id}:
    #Получение модели пользователя по ID
    get:
      tags:
        - users
      operationId: getUserById
      summary: Getting the user's model by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          description: 'User ID.'
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User model
          content:
            application-json:
              schema:
                $ref: '#/components/schemas/user'
        '400':
          description: Bad Reaquest.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
...
components:
  schemas:
    credentials:
      type: object
      properties:
        username:
          description: 'Username for auth or registration'
          type: string
          example: 'username@email.com'
        password:
          description: 'Password without encoding'
          type: string
          example: 'password'
...
    user:
      allOf:
        - $ref: '#/components/schemas/userIdModel'
        - type: object
          properties:
            username:
              description: 'Username.'
              type: string
              example: 'blackIIIfox@gmail.com'
        - $ref: '#/components/schemas/userModel'
        - type: object
          properties:
            avatar:
              description: 'User avatar file id'
              type: integer
              example: '1'
    userCommonInfoWithoutRoles:
      type: object
      properties:
        name:
          description: 'User is full name'
          type: string
          example: 'Rozkin Pavel Aleksandrovich'
        address:
          description: 'User is residential address.'
          type: string
          example: 'Russia, Moscow'
    userCommonInfo:
      allOf:
        - $ref: '#/components/schemas/userCommonInfoWithoutRoles'
        - type: object
          properties:
            roles:
              type: array
              items:
                $ref: '#/components/schemas/role'
    userCommonInfoWithBlocked:
      allOf:
        - $ref: '#/components/schemas/userCommonInfo'
        - type: object
          properties:
            isBlocked:
              description: 'The user is status, whether blocked or not.'
              type: boolean
              example: false
...
    userCreatableModel:
      allOf:
        - $ref: '#/components/schemas/credentials'
        - $ref: '#/components/schemas/userCommonInfoWithoutRoles'
    userUpdatableModel:
      allOf:
        - $ref: '#/components/schemas/userCommonInfoWithBlocked'
    userIdModel:
      type: object
      properties:
        id:
          description: 'Unique identifier of the user.'
          type: integer
          example: 12
    userModel:
      allOf:
        - $ref: '#/components/schemas/userCommonInfoWithBlocked'
...
    role:
      description: 'User role in system. Define user privileges in the system.'
      type: string
      enum:
        - Admin
        - Adventurer
        - Customer
        - Registrar
        - Distributor
      example:
        Admin
...
    errorModel:
      type: object
      properties:
        message:
          type: string
          description: 'Error description.'
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes